combine token manager example and its run script
